% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{COPASutils}
\alias{COPASutils}
\alias{COPASutils-package}
\title{COPASutils - Easy reading, processing, and manipution of COPAS BIOSORT data.}
\description{
COPASutils - Easy reading, processing, and manipution of COPAS BIOSORT data.
}

