% Generated by roxygen2 (4.0.1): do not edit by hand
\name{edgeEffect}
\alias{edgeEffect}
\title{Detect edge effects on 96-well plates}
\usage{
edgeEffect(plate, trait = NULL)
}
\arguments{
\item{plate}{a summarized and filled plate data frame}

\item{trait}{a singular trait to test, defaults to NULL and will test all traits}
}
\description{
Test for an effect of the position of wells in a 96 well plate. This function will split a plate population by edge wells and center wells and test the two populations for significant differences in either a specific trait or all traits if a trait is not specified.
}

