% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fillWells}
\alias{fillWells}
\title{Fill in any missing well rows with NA values}
\usage{
fillWells(plate)
}
\arguments{
\item{plate}{the data frame of the plate to filled}
}
\description{
Returns a data frame with any missing wells filled in as NA for all measured data
}

