% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCompare}
\alias{plotCompare}
\title{Visualize and compare values and distributions across multiple plates}
\usage{
plotCompare(plates, trait, plateNames = NULL)
}
\arguments{
\item{plates}{the list of plate data frames to compare}

\item{trait}{the trait to compare, entered as a string}

\item{plateNames}{an optional character vector with the names of the individual plates; if no names are entered, numbers will be used in the order the data frames are entered}
}
\description{
Plot the value (bar plot, if summarized) or distribution (boxplot, if unsummarized) of the data from each well across multiple plates.
}

