% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDR}
\alias{plotDR}
\title{Plot a dose response curve by strain across a plate}
\usage{
plotDR(plate, dosages, trait = "n")
}
\arguments{
\item{plate}{the plate data frame to be plotted}

\item{dosages}{a vector of dosages in the plate, entered by row}

\item{trait}{the trait to be plotted on the y axis}
}
\description{
Return a ggplot2 object of a dose response curve by strain across a plate
}

