% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDR_allTraits}
\alias{plotDR_allTraits}
\title{Plot dose response curves for all traits}
\usage{
plotDR_allTraits(plate, dosages)
}
\arguments{
\item{plate}{the plate data frame to be plotted}

\item{dosages}{a vector of dosages in the plate, entered by row}
}
\description{
Return a list of ggplot2 objects of dose response curves by strain across a plate. Plots for specific traits can be accessed by name from the returned list (i.e. "plots$n" will return the dose response plot for the trait "n").
}

