% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{readSorter}
\alias{readSorter}
\title{Read in raw sorter data}
\usage{
readSorter(file, tofmin = 0, tofmax = 10000, extmin = 0, extmax = 10000)
}
\arguments{
\item{file}{path to sorter data file}

\item{tofmin}{minimum cut off for time of flight, defaults to 0}

\item{tofmax}{maximum cut off for time of flight, defaults to 10000}

\item{extmin}{minimum cut off for extinction, defaults to 0}

\item{extmax}{maximum cut off for extinction, defaults to 10000}
}
\description{
Reads a raw sorter file into a dataframe, removing NA values and any objects not fitting in to the min and max cut offs
}
\examples{
readSorter("SortTest.txt", 60, 2000, 60, 5000)
readSorter("SortTest.txt", tofmin=60, extmin=60)
}

