% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{removeWells}
\alias{removeWells}
\title{Remove wells from a data frame representing a plate}
\usage{
removeWells(plate, badWells, drop = FALSE)
}
\arguments{
\item{plate}{a processed data frame that has been run through processWells}

\item{badWells}{a character vector consisting of all wells to remove}

\item{drop}{a logical value dictating whether to drop wells from the data frame or set measured values to NA, defaults to FALSE}
}
\description{
Returns a data frame representing a plate with bad wells removed (either with phenotype variables set to NA or rows dropped entirely).
}
\examples{
removeWells(processedPlate, c("A1", "B7", "F12")) #Sets phenotype values for wells A1, B7, and F12 to NA
removeWells(processedPlate, c("A1", "B7", "F12"), drop=TRUE) #Removes rows corresponding to wells A1, B7, and F12
}

