% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{summarizePlate}
\alias{summarizePlate}
\title{Condense all objects examined to appropriate well}
\usage{
summarizePlate(plate, strains = NULL, quantiles = FALSE, log = FALSE)
}
\arguments{
\item{plate}{plate to summarize, must be run through readSorterData, not readSorter}

\item{strains}{a vector of all strain (sample) names input row-wise to add to the data frame}

\item{quantiles}{if TRUE, columns of trait quantiles (every fifth) will be added to the output dataframe, defaults to FALSE}

\item{log}{if TRUE, columns of log transformed EXT and red fluorescence will be added to output dataframe, defaults to FALSE}
}
\description{
Returns a data frame with one row per well with summary statistics.
}
\examples{
processWells(plate, quantiles=TRUE)
processWells(plate, quantiles=TRUE, log=TRUE)
}

