% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPASutils-package.r
\docType{package}
\name{COPASutils}
\alias{COPASutils}
\alias{COPASutils-package}
\title{COPASutils - Easy reading, processing, and manipulation of COPAS BIOSORT data.}
\description{
COPASutils - Easy reading, processing, and manipulation of COPAS BIOSORT data.
}

