% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractTime}
\alias{extractTime}
\title{Set time to relative}
\usage{
extractTime(plate)
}
\arguments{
\item{plate}{a plate to extract the time from}
}
\description{
Sets time relative to first well run, used in other functions, not meant to be used on its own.
}

