% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCorMatrix}
\alias{plotCorMatrix}
\title{Plot a correlation matrix within a plate or between plates}
\usage{
plotCorMatrix(plate1, plate2 = plate1)
}
\arguments{
\item{plate1}{one plate to compare in the correlation matrix}

\item{plate2}{an optional plate to compare plate1 to, defaults to plate1, if no argument is entered plate1 will be compared to itself}
}
\description{
Returns a ggplot2 object of a correlation plot for all traits within a single plate or between two plates
}

