% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCorMatrix}
\alias{plotCorMatrix}
\title{Plot a correlation matrix within a plate or between plates}
\usage{
plotCorMatrix(plate_summary1, plate_summary2 = plate_summary1)
}
\arguments{
\item{plate_summary1}{one summarized plate to compare in the correlation matrix}

\item{plate_summary2}{an optional summarized plate to compare with plate_summary1, defaults to plate_summary1; if no argument is entered plate_summary1 will be compared to itself}
}
\description{
Returns a ggplot2 object of a correlation plot for all traits within a single plate or between two plates. The data from each plate must be summarized prior to being passed to this function.
}
\examples{
#### COPASutils Figures
### Figure 4
# library(dplyr)
# sumDose <- summarizePlate(doseData)
# corDose <- select(sumDose, -n.sorted)
# plotCorMatrix(corDose)
}

