% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDR}
\alias{plotDR}
\title{Plot a dose response curve by strain across a plate}
\usage{
plotDR(plate_summary, dosages, trait = "n")
}
\arguments{
\item{plate_summary}{the summarized plate data frame to be plotted}

\item{dosages}{a vector of dosages in the plate, entered by row}

\item{trait}{the trait to be plotted on the y axis}
}
\description{
Return a ggplot2 object of a dose response curve by strain across a plate
}
\examples{
#### COPASutils figure
### Figure 5
# strains <- rep(c("A", "B", "C", "D"), each=6, times=4)
# sumDose <- summarizePlate(doseData, strains=strains)
# dose <- rep(c(1, 5, 10, 15, 20, NA), times=16)
# plotDR(sumDose, "n", dosages=dose)
}

