% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readPlate}
\alias{readPlate}
\title{Read in the sorter data with minimal processing}
\usage{
readPlate(file, tofmin = 0, tofmax = 10000, extmin = 0, extmax = 10000,
  SVM = TRUE, reflx = TRUE)
}
\arguments{
\item{file}{path to sorter data file}

\item{tofmin}{minimum cut off for time of flight, defaults to 0}

\item{tofmax}{maximum cut off for time of flight, defaults to 10000}

\item{extmin}{minimum cut off for extinction, defaults to 0}

\item{extmax}{maximum cut off for extinction, defaults to 10000}

\item{SVM}{logical dictating whether to predict bubbles with the SVM}

\item{reflx}{logical indicating whether ReFLx module was used (TRUE) or LP Sampler was used (FALSE), defaults to TRUE}
}
\description{
Returns a minimally processed data frame, optionally with SVM-mediated bubble prediction.
}

