% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizePlate}
\alias{summarizePlate}
\title{Condense all objects examined to appropriate well}
\usage{
summarizePlate(plate, strains = NULL, quantiles = FALSE, log = FALSE,
  ends = FALSE)
}
\arguments{
\item{plate}{plate to summarize, must be run through readSorterData, not readSorter}

\item{strains}{a vector of all strain (sample) names input row-wise to add to the data frame}

\item{quantiles}{if TRUE, columns of trait quantiles (every fifth) will be added to the output dataframe, defaults to FALSE}

\item{log}{if TRUE, columns of log transformed EXT and red fluorescence will be added to output dataframe, defaults to FALSE}

\item{ends}{if TRUE, columns of min and max values for all traits will be added to output dataframe, defaults to FALSE}
}
\description{
Returns a data frame with one row per well with summary statistics.
}
\examples{
# exampleStrains <- rep(c("N2", NA), times=48)
# plate1 <- summarizePlate(plateData1, quantiles=TRUE, log=TRUE, ends=TRUE)
# plate2 <- summarizePlate(plateData2, strains=exampleStrains)
}

